+++ did/shared/src/main/scala/fmgp/did/comm/OpaqueTypes.scala
@@ -87,71 +87,3 @@ object MsgID:
   extension (messageID: MsgID) def value: String = messageID
   given decoder: JsonDecoder[MsgID] = JsonDecoder.string.map(MsgID(_))
   given encoder: JsonEncoder[MsgID] = JsonEncoder.string.contramap[MsgID](_.value)
-
-opaque type Payload = Base64
-object Payload:
-  def fromBase64url(data: String): Payload = Base64.fromBase64url(data)
-  extension (data: Payload)
-    /** decode the base64url to a string */
-    def content: String = data.decodeToString
-    def base64url: String = data.urlBase64
-    def base64: Base64 = data
-  given decoder: JsonDecoder[Payload] = Base64.decoder
-  given encoder: JsonEncoder[Payload] = Base64.encoder
-
-/** JWM_SIGNATURE is a Base64 url encode */
-opaque type SignatureJWM = String
-object SignatureJWM:
-  def apply(value: String): SignatureJWM = value
-  extension (signature: SignatureJWM)
-    def value: String = signature
-    def base64: Base64 = Base64.fromBase64url(signature.value)
-
-  given decoder: JsonDecoder[SignatureJWM] = JsonDecoder.string.map(SignatureJWM(_))
-  given encoder: JsonEncoder[SignatureJWM] = JsonEncoder.string.contramap[SignatureJWM](_.value)
-
-opaque type ProtectedHeaderJWT = Base64
-object ProtectedHeaderJWT {
-  def fromBase64url(data: String): Payload = Base64.fromBase64url(data)
-  extension (header: ProtectedHeaderJWT)
-    /** decode the base64url to a string */
-    def content: String = header.decodeToString
-    def base64url: String = header.urlBase64
-    def base64: Base64 = header
-}
-
-opaque type SignatureJWT = Base64
-object SignatureJWT {
-  def fromBase64url(data: String): Payload = Base64.fromBase64url(data)
-  extension (signature: SignatureJWT)
-    /** decode the base64url to a string */
-    def content: String = signature.decodeToString
-    def base64url: String = signature.urlBase64
-    def base64: Base64 = signature
-
-}
-
-opaque type JWT = (ProtectedHeaderJWT, Payload, SignatureJWT)
-object JWT {
-
-  def apply(protectedHeader: ProtectedHeaderJWT, payload: Payload, signature: SignatureJWT): JWT =
-    (protectedHeader, payload, signature)
-
-  def fromStrings(protectedHeader: String, payload: String, signature: String): JWT =
-    apply(
-      ProtectedHeaderJWT.fromBase64url(protectedHeader),
-      Payload.fromBase64url(payload),
-      SignatureJWT.fromBase64url(signature)
-    )
-
-  def unsafeFromEncodedJWT(str: String): JWT = str.split('.') match {
-    case Array(protectedHeader: String, payload: String, signature: String) =>
-      fromStrings(protectedHeader, payload, signature)
-  }
-  extension (jwt: JWT)
-    inline def protectedHeader: ProtectedHeaderJWT = jwt._1
-    inline def payload: Payload = jwt._2
-    inline def signature: SignatureJWT = jwt._3
-    def base64JWTFormatWithNoSignature = jwt.protectedHeader.urlBase64 + "." + jwt.payload.urlBase64
-    // def base64: Base64 = Base64.fromBase64url(jwt.value)
-}
